Add some description in the protocol header
Since we're mostly adding formatting descriptions in comments next
to the individual fields, we'll add some description for the
username and charname fields - they should both be in UTF-8.  We
won't worry about the password field, since it'll be going away
at some point (hopefully) soon.

While we're at it, we'll add a short note for the LGTREQ packet
type in the preamble comment.