Add protected blank constructor for client comm
There is a lot of mocking that needs to go on in order to test the
client comm object, most of which involves making a fake addrinfo
structure, and in most cases it is totally unnecessary.

We'll add a protected blank constructor so it can't normally be
called, but can be made public when we derive our fake object.
We'll move all the initialization into the init() method, also
protected, so we can call it in both constructors.