Add a guard for started threads in client comm
When doing a thread cleanup on destruction, we can't know whether
the threads have been started or not, so we can't really expect the
various calls to succeed - there may not be threads to cancel and
join.  The server's basesock object has a 'thread_started' flag
which keeps track of whether the threads are actually started, and
the stop methods work or don't work depending on that flag.  We'll
do the same thing here.

The comm exception test needs a little help in the happy-path test,
in that by that point, we haven't successfully started the threads
yet.  All the error flags are off, so we can do a regular startup
before we do the delete.