Add exceptions to client comm shutdown
We make a bunch of library calls which may fail, and don't check
any of the return values.  We'll throw exceptions as we typically
do in these situations - message, error string, return value.

In the destructor, when we do call the stop() method, we'll catch
any stray exceptions and log them.