Add mocking and testing for failed malloc in dh exception test
OPENSSL_malloc, which is what the dh_shared_secret function calls,
is actually a macro around CRYPTO_malloc, which we need to mock.
The test itself is pretty straightforward.

OpenSSL changed the signature of CRYPTO_malloc from 1.0.x to 1.1.x
by changing the size parameter from int to size_t, so we do need
to take that change into account.