Add test for proto dh key context failure
When we create a shared secret, we must create a key context.  If
it fails, that will drop us into our first bailout condition, which
requires us to mock the OPENSSL_free function.  It's a macro around
CRYPTO_free, so that's our actual mock.  The signature of
CRYPTO_free changes from 1.0.x to 1.1.x, so we need to take the
differences into account.

We'll also add a counter to make sure we're freeing appropriately.
In some other bailout conditions, we do call it once more, so this
will help us with those tests.