Add proto test for dh derivation failures
We initialize the derivation context, set the peer key, and perform
the derivation.  If one of those fails, it drops us into the second
of our bailout conditions.  This requires us to mock
EVP_PKEY_CTX_free.  We add counting in the new mock to be sure our
mock is being called.