Test first two bailout conditions in proto digest
When calculating a digest of a message, we need to test the various
bailout-on-error conditions.  The first is a direct return, but the
next one drops down to the bailout section.

CRYPTO_malloc is already mocked for us by the previous set of
tests, so fortunately it's a simple couple of tests.