Add free_dh_message function
Instead of making people free both allocated parts of the
dh_message structure we return, we'll have a convenience function
that'll do it for them.

We can also use it inside the dh_shared_secret() function.  Once
we have taken the digest of the actual shared secret, we can blow
away that shared secret in one call.