Add pubkey data to database authentication methods
When authenticating users, we will eventually replace the password
with the public key we receive.  We'll have it in the database,
which we'll check instead of the password.

For now, we'll add the key and the length of the provided key to
the method signatures everywhere, but not do anything with them
yet.

Re:  issue #32