Add proto function to extract public key from EVP_PKEY
EC keys have the ability to extract the public key from the private
key.  We do eventually need to send it to the other end, so we
need to be able to get it into a buffer somehow.  We'll add a
simple function to take care of it for us.