Add player_keys table to SQL description files
Since we need to keep track of player public keys, we need a table
in our database to record them.

While we were at it, we noticed that the timestamp fields in the
player_logins table were using the MySQL-specific DATETIME data
type.  The SQL standard type is TIMESTAMP, so we'll switch these
fields to use that type instead.