Check public keys on login to the server
The new login packet includes a public key in addition to the
password field, so we'll alter the queries and logic to check the
key we got against the latest key in the database.

Since the client does not yet send the key, this will render all
clients unable to login until that is resolved.

Re:  issue #32