Actually send the server key during login
The protocol needs us to send the key once a connection is made.
We do some authentication checking before we send anything, and
send the key before we do any actual access checking.

The test is simple enough - we should have sent two packets by the
time a user is fully logged-in.

Re:  issue #32