Add IV to server key packet
The two endpoints of an encryption that we're trying to do need to
share an initialization vector, which needn't be secret.  We'll add
that to the server key packet, so the client will get everything
it needs at the same time.

Also while we're at it, we'll correct the size of the sequence
field in the server key packet; we added it after we made the
initial sequence size change, which we later reverted.  When we
did the revert, the server key packet didn't have anything to
revert, so it's been wrong since then.