Add sequence to IV before encrypt/decrypt
We've always intended to use the sequence field as part of the IV
when performing encryption or decryption operations.  When
inserting the sequence, a 64-bit int, into the IV, an array of
unsigned bytes, we can't count on the byte-ordering of the integer
to be consistent from platform to platform.  We'll convert it to
network byte order as we insert it into the IV.

This requires the encrypt and decrypt functions to alter the IV,
so we need to drop the const specifier.

We do this in the proto library rather than in the server and
client to simplify the calling code; both ends will need to be
doing the same things, so we'll make sure they're always done.

Re:  issues #32, #33