Eliminate the default optimizations for compilers
The AC_PROG_CXX macro will set a default CXXFLAGS of '-g -O2' if
there is not already one set; likewise for AC_PROG_CC and CFLAGS.
Once we stopped needing a CXXFLAGS passed in on the command line,
the test suite stopped working due to some of our test conditions
being optimized out.  We'll default to empty CXXFLAGS and CFLAGS
variables, to prevent confusion and breakage.