Add simpler base_user constructor
In order to simplify mocking in our test suite, we'll add an
almost-default constructor, which blanks pretty much every data
member.  It's protected, so it'll only be usable if the class is
derived, say, for purposes of mocking.

Also add a fake base_user header for the test suite, and start
using it where possible.

We're about to start making significant changes to the base_user,
particularly in the constructor, so having something simpler to
instantiate will be a big help to our test suite.