Move most of the base_user initialization into the constructor
We've been doing most of the initialization of the base_user in
methods of the listen_socket, rather than in the base_user
constructor where they belong.

There are huge changes to the test suite for both the base_user and
the listen_socket.  This should reduce the coupling between the two
objects, which is always a good thing.

This is a big commit; breaking it up into smaller changes would
have been extremely complex.