Add method to calculate shared secret to base_user
Once we figure out that the user is who we think they are, we need
to add the shared key into the base_user object.  The listen_socket
doesn't need to know how it's done, because it'll be relying on the
user itself to do the encryption and decryption.

Re:  issue #32