Return success/failure from encrypt/decrypt base_user methods
Once we start using the encrypt_packet and decrypt_packet methods
in the socket packet handlers, we'll need some way for the callers
to know whether the operations have succeeded.  Exceptions would
be an extremely clunky way to go here, so we'll return nonzero on
success, and zero on failure.