Take local copies of the IV during encryption/decryption
There was a race condition with modifying the user's encryption IV
directly, in that we're copying the sequence number over part of
that set of bytes without any locking.

To both eliminate the race condition, and eliminate the need for
locking, we'll take a local copy of the IV within both the
encryption and decryption functions, and modify that copy.