Repair sizing bug in client comm
When we test a newly-received packet as to whether it is a "good"
packet, we test against our array of packet types.  Taking the
sizeof that array takes the size of the whole thing, rather than
giving us the number of entries.  Taking the size of the whole
thing and scaling by the size of each entry gives us the size we
actually need.

While we're at it, we'll repair the error message we emit for this
failure - the packet type is basically a char, and is being
interpreted by the log stream as a char, where we want an integer.