Make client comm bad-packet flag do something
The client comm test has a bunch of mocked-out functions, which we
control with a set of flags.  We've got a flag 'bad_packet', that
has never been hooked up to do anything.  When that flag is set,
we'll have the fake recvfrom function return a packet with an
invalid packet type.