Pull in the client encryption key when we get the password
We're switching away from using the password as an authentication
mechanism with the server, and towards it being a passphrase for
the client encryption key stored on disk.

We still use it for both, but soon enough we'll drop the auth
usage, and only use it as a local passphrase.

Re:  issue #33