Correct sensing missing conf file in client
Opening a missing file in a stream doesn't throw an exception like
we had initially coded, so our sensing that there was a missing
file wasn't doing anything.  The constructor will return a not-
object, which we can sense by taking the negation of it.

This allows us to drop the configure checking for the exception
element that we were trying to use, but may or may not have had.