Actually send the public key from the client
When we loaded the private key into the config object in the client
we never actually converted to the public string, and we never
actually copied the public string into the login packet.  Now that
we're doing both, login should be much more effective.

While we're at this, we'll fix up an include or two.

Re:  issue #33