Repair server socket login logic
We had "streamlined" the logic in accepting packets in both the
server socket types, but that assumed that all packets were coming
in from logged-in users; nobody could login, because they were not
found in the objects' descriptor lists.

We'll allow non-logged-in users to proceed through the process,
and just send NULL in place of the user type.