Repair protocol packet_size function
The packet_size funtion is partly a guardian of what is a proper
packet, and the first test is to make sure the packet type doesn't
go any higher than the types we've defined.  We have added two
types since we last looked, so the logout request and server key
were being declared as invalid by this check.  This needs a better
mechanism to figure out what the highest-numbered packet type is.