Move creation of user login log entry into base user
Having something that isn't base_user digging into base_user to
make a string seems to increase coupling unnecessarily.  We'll add
a simple method to have the base_user spit out a string that
describes itself, for this same purpose.  This simplifies the
login method within the listen_socket a whole lot.  It could end
up being useful in other contexts in the future.

This method isn't exciting enough to do a full-up test, so we'll
add a couple diag() calls into our existing tests, to exercise both
of our possible branches (with and without default_slave).