Repair incorrect MySQL check_authentication logic
When we authenticate a user, we take a username and a public key
and see if they compare to what's in our database.  Doing compares
against BLOB columns in MySQL needs a bit more hand-holding than a
typical string or integer column.  We convert the key we get into
a hex string, and then use the MySQL HEX() function to compare what
we converted against what we have stored.