Repair client comm test of bad encryption
When the client comm object can't encrypt an outgoing packet, there
should be a specific log message.  However, we had chosen a packet
type which always "succeeds" encryption, because nothing in it
actually gets encrypted (it's a basic packet, and none of the basic
fields get encrypted).  Choosing a packet that's bigger than a
basic packet makes the test work as we expect.