Streamline path evaluation in client config
We don't need to go through the secondary step of turning a char
array into a std::string, in order to do the replace for home
directories.  We can use the result of getenv directly.

While we're at it, we'll improve the test coverage here, making
sure that our throw is working as expected.