Generate a public key file when generating a client key
We've got to get our clients' public keys to the server somehow,
so we can insert them into the database.  It'll be easier to do it
if we've got the data in a file somewhere, so we'll go ahead and
generate one while we're doing the private key.

No mechanism to send that public key anywhere yet, but at least
we're ready for it when it arrives.