Improve log messages from zone
When we start up, we load objects, but we never know how many we've
actually loaded.  Wonder no more!

As we shut down, if we do delete objects, we'll tell how many we're
going to delete.  If the number is zero, we won't say anything.
Also, while we're in the area, we'll turn our deletion loop into a
range format loop.  They're so much more readable than classic for
loops.