Remove open/check/close user login database methods
We no longer want to maintain user login information in the
database in the way that we had coded.  We weren't calling the
methods, and they wouldn't have worked very well anyway, so for
the time being, we'll get rid of them.

We'll revisit how we want to do user logging at later time.

Re:  issue #208