Drop the SEGV handler
Initially, we thought that having a stack trace directly in our log
would be super-helpful, but the backtrace that we get is not very
useful.  It would be much more useful to have the program dump core
so that we can get it into a debugger.  We've tried a couple
different methods to get the program to dump core, but the only way
that really works is to use the default SEGV handler.

As a result of that, we'll dump our SEGV handler.  It's the only
place in the codebase which includes the execinfo.h file, and the
only place which calls backtrace(), both of which are specifically
checked for in the configure file, so we can drop both of those
checks as well.