Fail configuring if needed headers/functions are missing
Until now, we've just been checking for the standard headers and
functions, and going by our defines whether they needed to be
included.  That makes the source files more verbose and difficult
to read than they need to be, so we'll shift the decision up into
the configure script - if the files aren't there, we'll fail the
configure step.

As a result, we can get rid of a LOT of #if/#endif within several
files around the codebase.

There are a small handful of files and functions which we do some
conditional compliation on, surrounding byte-swapping and byte
ordering, so we'll separate those out into their own non-failing
checks.