Convert MySQL authentication to use prepared statement
In order to limit SQL injections against our queries, we'll convert
all of them to use prepared statements; the DBMSs have a much
clearer understanding of how strings need to be escaped for their
own uses than we do.  For this particular query, it allows us to
pass in the public key itself, rather than jumping through hoops
to manually format it in hex and adjust the query to force the type
into the same hex format.

We want to be able to have consistent queries between all database
types, so we'll insert them into the main DB class.  Currently for
the PGsql query we're using current_timestamp, but NOW() should
work fine in that context, so we should not need to override the
query string.

Re:  issue #207