Convert MySQL authorization query to use prepared statement
The first of the authorization queries, which takes a userid and
characterid, is the next to be converted.

Until now, we've been inexplicably using the server IP address as
the value to figure out what's allowed on our server.  We already
have the serverid, so let's use it instead to reduce the number of
tables we're joining.

Re:  issue #207