Convert MySQL player server skills to prepared statement
Last query to be converted.  As with most of the other queries,
we also convert from using the host IP to the hostid to reduce
the number of tables in the join.

Our table fetches data as MYSQL_TIME, which is close but not
identical to time_t, so we need to add a conversion routine.

Since this removes the last of the sprintf calls in the file, we
no longer need the PRI* formatting macros, and we can drop the
inttypes.h include.

Re:  issue #207