Add geometries table to MySQL SQL file
Eventually we'll need to have a record of geometry-related data,
so we'll start out a new table which we can reference from the
server_objects table.  We don't have anything of consequence in the
new table yet, but we can add things later, once we know what we
want to add.

Also a slight reordering of the tables; makes a little more sense
with the server-related things mostly together.