Add SQL file for PgSQL
We've refined our MySQL SQL file to a state which is as up to date
as we can manage, so we can copy and convert to a format that
PostgreSQL can use.  The formats are close, but not directly
compatible; some of the MySQL integer types are not available in
PgSQL, and the table constraint syntaxes are different.  The BLOB
types are also different.

Now that we have a PostgreSQL that we can test against, we can
verify that the commands are indeed correct.  Verified against
PostgreSQL 9.6.