Retrieve server ID in PgSQL object
A number of our queries could be optimized if we had the server's
ID available, instead of having to join yet another table via the
server's IP address.  The base DB class has a host_id member, which
we'll populate as needed when we first connect to the database.

It turns out that using PQexecParams is a bit simpler than making
a prepared statment and executing it every time.  It takes about
the same parameters, and we get the same benefits, without the
extra call.

Since we're going to be converting everyting to use parameterized
statements, we'll use one to do this query.

Re:  issue #207