Import latest CuddlyGL changes
There have been significant changes to the CuddlyGL codebase over
the last while, which cascade down into a lot of our client code.
All constructors now have varargs formats, and the canned x/y
sizes are gone.  We can combine all of the constructor calls with
the set calls that usually follow.

Recently, the facility for setting negative positions has been
added, which removes the need for a lot of our log_display's custom
(and not-fully-functional) positioning code.

We had a couple of defines in the log_display, one of which we were
able to remove entirely due to the negative offset changes.  We
converted the other to a const class variable.