Repair incorrect GLM headers
GLM has moved the quaternion headers around a bit since earlier
versions, and we've been including the wrong files.  Correct files
are in the gtc/ (core?) directory, rather than gtx/ (extensions).
We do need to change the name of one instance of fquat to quat,
in the client object structure; seems like the float version is
basically the default, so probably no need to give it a funky
name.

The CuddlyGL library was also using some semi-internal headers
which have been renamed.