Add a virtual base class for control devices
In order to have a consistent interface with any control device, we
should have some sort of base object which defines an interface for
any control device we might have.

We may have callbacks that we want to add, or we may want to deal
directly with a comm object.  From the interface, we can't necessarily
predict which one a given control object would want, so we'll always
take both as arguments, so the interface is consistent.