Convert game object rotation to a quaternion
We can't properly represent multiple rotations with the single
vector we've had in the game object.  We'll change it over to a
quaternion, which will allow us to represent as complex a rotation
as we might ever need.

We'll also add a method to the game object to do scaled movement
and rotation. We call the scaling factor 'interval', since it
represents the time interval in seconds since the last motion.
This removes any need to know anything about how motion or
rotation works from the motion pool workers - they just tell the
objects to do their thing, and the objects do it.

While we were in the motion pool, we needed to change the "are we
still moving" checks.  It seemed like a fine time to simplify as
well.  If we keep null-motion and null-rotation objects around, we
can compare to them directly, and have much clearer logic.

We also need to update the stop action routine, since it does deal
with a game object's rotation.