Add some logging on ack of action request
In the client, we've been logging ack packets, just to make sure
our traffic is getting to the server and back.  We didn't have any
specific clause for an ack of an action request, so they were all
being logged as unknown acks.  We'll add a simple clause for them,
and a quick test for the new clause.