Reorder quaternion field handling
We've been putting the w element of the quaternion last, but we
should order them in alpha order.  It's not entirely important what
order we actually use, since we do all our conversions from internal
types to packet types field by field, but for consistency's sake,
we'll handle things in the order they're now specified in the
structure.