Have screen follow the user's object
For our first trick, we'll get the client camera to follow the
user's object, 5m behind and 1m above.  We still can't turn, but we
can see our default objects moving around as we move.

In order for our math not to be crazy, we'll normalize each
object's look vector as it arrives.

Re:  issue #217