Make action request target point signed
We've been trying to get backward movement working, and the problem
was that we couldn't pass negative numbers in any of the vector
elements.  Making the packet elements signed allows us to pass in
negative values, and allows backward movement.

We add a few lines to handle backward movement in the client key
callback routine.

We'll also update the docs for this field.