Add scaling constant for action request position fields
We scale all position fields in the action request packet by a
given amount, 100.  So that everybody understands that scaling
factor, and can have a name to give to the magical constant, we'll
set a define for everyone to use.

We also noticed that we weren't doing any scaling in the server.
Shame on us.