Move game object skills struct into game object
One of the items within defs.h is a structure which is used within
the game object, to keep track of the player's skills.  We'll move
it into the place where it's used, and out of the random catch-all
file.  While we're at it, we'll make the map into an unordered_map,
which has better average access time.

The names of the types cascade down into the database objects, as
they feed values into these structures.