Move sequence definition to geometry object
We don't actually use the sequences at the moment, since they are
ill-defined for what we want to do, but having the definition of
the structure in a different place than the structure is actually
(not) used is inappropriate.  We'll move it where it belongs, and
comment the whole lot of it, since it's unused.