Add game object map typedefs to game object
We use maps of game objects in a number of places throughout the
server codebase, and we always need to use the FULL declaration for
them.  This makes it impossible to, say, change the type of the
object if we wanted to.

So we'll add some typedefs into the GameObject class for maps and
iterators into those maps.

While we're at it, we do want to change the maps to use the new
unordered_map type, since element access is O(1) on average.