Convert game object maps/sets to unordered
We've got a couple of maps/sets in the GameObject, which keep track
of ancillary values - attributes and natures.  Switching over to
unordered versions of the same data structures will give us average
O(1) element access time, instead of O(log n) for the ordered
versions of the same data structures.