Add header for action class
defs.h has an action_rec class which handles server actions, their
levels, names, etc.  We want to get rid of defs.h, because it's a
catch-all, and (IMNSHO) terrible practice.

We'll add a header with this class, and rename it to Action.  We'll
make the changes to the rest of the codebase in another commit; for
now, this is the class, and it's not hooked up to anything else.