Move interval calculation inside game object
When moving an object, the game object knows when it was last
updated, and it knows how much it should be moving, so it can
figure out the interval and amount it needs to move.  Since we no
longer have to do any of this inside the motion pool worker, it's
much simpler.