Move activation/deactivation of game object into game object
We've got places which add and remove natures from a given game
object spread all over the server, which is inappropriate.  We'll
move the standard "activate" and "deactivate" actions into the
game object itself, where it can know the right things to add and
take away, and they can be done in close proximity, so if we need
to change things in the future, we don't have to go far.

These changes cascade into the database objects, which create
character objects as deactivated.