Make game object natures an enum
Instead of using strings for our natures, which take a while to
compare, we'll use an enum, which is just an integer, and compares
very quickly.

We lose the open-ended facet of the way we've done natures to now,
but at this time the performance gain will outweigh the convenience
loss.