Add standard movement factors to server move actions
Standard walking speed for humans is around 1.4 m/s, so we'll add a
factor define which turns this into a constant.  We don't have a
good metric for "how fast do people turn", so 1 radian/s is
probably a fine place to start.  It seems fine in practice.

While we're here, we'll do some very minor cleanups in the comment
header.