Turn object to face proper orientation
Now that we have good orientations for each object, we should use
that orientation when setting the model matrix for them.  After a
lot of trial, error, and research, it's as simple as multiplying
the result of the translation by the orientation; I was trying to
do it the wrong way around, with terrible results.

Re:  issue #217