Add accessors to the game object
We get a couple of benefits out of having the get/set operations
inside the game object.  First, this will allow us to remove the
recordkeeping we do (timestamp update) from all over the codebase,
and segregate it inside a thing that actually cares about the
timestamp.  Second, we are likely to need locking on these items
at some point in the future, and having all the action handled
inside the objects themselves eases the transition to locking.

There's a looooot of other changes around the codebase that we have
to remove.  We'll do that in other commits.