Update client includes for general Cuddly use
We've been depending on our bundled version of CuddlyGL, which
limits us to requiring *only* the version that we provide.  Cuddly
is a separate project, and moves at its own pace.  In general, the
interface will be mostly compatible, so we shouldn't need to depend
on any particular version.

For now, we'll require the includes from the bundled version with
'-I.' in the makefile.  In the future, we'll add some autoconf
pieces, and properly disconnect things.