Enable X11 key repeat handling in Cuddly connector
We're seeing a lot of poor behaviour and segfaulting in both server
and client, and it seems like part of the reason is that we're getting
bad repeat behaviour.  There's a server side problem too, but we'll
start with the client side.  CuddlyGL's connector adds handling of the
XAutoRepeatOn/XAutoRepeatOff during focus changes, to make the
in-library repeat work correctly.

In order to set the repeat back to what it was before we started, we
need to call the disconnect function as we exit the client program.