Bail out of direction key handling if not connected
In the client, we have a key handler in the main window, which takes
arrow key input, and turns it into action requests.  We figure out the
right request to make before we bother figuring out if we're even
connected, which is pointless.

We'll bail out as soon as we can, if there is no functioning comm
object in our array.