Allow 0-values to stop a given movement
Previously, when we received 0-length vectors for either movement or
rotation from the client, we'd drop them.  The only way to stop any
movement or rotation was to stop *all* movement *and* rotation via the
'stop' action.

Instead, we'll allow a continuing movement or rotation to remain,
while stopping the other, by using a 0-vector to do the partial
stopping.