Add locking in game object
There could be many things trying to operate on the game object at any
time.  We'll implement a read-write lock with two types of locking
methods.

Most systems that I looked at have different schemes for handling lock
contention.  Linux favors reads over writes, but whenever there is a
write pending, it's because of a request by a user, or some outside
influence, so it should be of a higher priority.  Linux headers define
PTHREAD_RWLOCK_PREFER_WRITER_NP (that last bit is "non-portable"),
which we'll try to set if it exists.  Other systems favor writes over
reads by default; MacOS seems to be one of these systems.

This commit merely adds the lock to the class and adds some locking
primitives.  We'll incorporate those into the various methods in a
future commit.