Use locking in game object
Now that we have locking primitives available in the game object,
we'll use them in the appropriate places:  read-locking in the
read-only methods, and write-locking in the setter methods.

In the general case, each of these methods should be able to be
treated as atomic operations; most of the logic involving the data
here will either be handled directly within the object, or there's a
whole lot more that we don't know about (gravity, collisions with
other objects, etc.) that must be handled elsewhere.  This is a start,
and may require some fine-tuning later on.