Make sure added objects are within the zone
When determining what sector an object should be in, we ought to first
make sure that it actually falls within the entirety of our zone.
Otherwise, we'll get a nice segfault.

If we find that we aren't in the zone, the motion pool should *not*
reinsert the object.  It'll just fall on the floor for now, but we'll
come back at some point, once neighbors are a thing.  Or once we get
collision, maybe we'll just collide with the edge.