Add locking to octree
We may run into instances in which multiple workers are trying to
update things within a single octree octant at the same time, so we do
need to put some locking in place.  We'll keep things private for now,
and only keep the locking within the octree methods; if, in the
future, we need to expand the critical portions (say, in the motion
pool, when something moves in a tree), we can certainly change the
availability.