Have server adjudicate orientation in movement
We've been having the client calculate its motion, and it makes
decoupling movement and rotation particularly knotty; the client would
have to decide whether its rotation should or should not influence its
movement.  Not what we want - the server should be the sole arbiter of
where the client's object goes.

To that end, the movement request will be based on the object
coordinates of the base object:  forward/backward is +/-y, and
left/right is -/+x.  The server game object, which is what handles the
movement, will then scale the movement of the object by its
orientation.  It's simple, and it works.

It's not completely clear how we should do multiple-axis movements or
rotations.  Do we want to have multiple requests, or should the client
combine requests and send just one?  For rotations, combining the
requests will require a protocol change, since the way we currently
specify rotation is with an axis, rather than a full quaternion.
We'll just have to try out the different schemes, and see what works
best.  Now that we have locking in the game objects, sending multiple
requests is not nearly so risky.