Add keyboard controller into client program
We'll use the keyboard controller in place of the existing callbacks.
We add the appropriate makefile pieces to allow the builds to occur.

For the time being we'll use only a single control type, but
eventually we'll have multiple control types, which are set via a
configuration option.  Each one of these types will eventually be
able to manage its own config parameters.