Only build shared loadable modules
There's no reason that we should continue building static versions of
the library files which we only intend to load dynamically, so we can
have libtool turn off the static versions by default.  If somebody
really wants to build them, they ought to be able to provide command
line options to configure that will turn them back on.

Also, we'll never be using the .la files for these loadable modules -
they're only used by the linker.  We can delete them from the install
directories, just to simplify what we're dropping on the filesystem.