Add lcov target to main Makefile
So that we don't have to remember the commands to generate a nice lcov
summary of test coverage, we'll add a Makefile target to run it for
us.

We'll add checks for the appropriate commands in configure.ac, and
have a conditional around the targets in the Makefile, whether we
found the appropriate programs or not.