Add client control factory
We'll want to be able to change control methods, based on the modules
which are present.  We'll shift all the control modules over to a
segregated directory, so nobody else will accidentally try to load
them.

We also make a factory class, which will let us know what modules are
available, and also load a given module for us and spit out an
instance of that control class.