Use control factory in main client program
Now that we have the control factory, we can insert it into the
appropriate spot in the main client.  We'll use a config item for
which control object we should use, so we'll add that element to the
client config object.  We default to the keyboard, since everybody
ought to have one.

Eventually, we'll add methods to the control objects to generate
CuddlyGL widgets for managing the various tweakable parameters for the
given controller.  For now, we just take whatever settings the given
controller already has.