Stop logging action and unknown ack packets
When doing a lot of movement, we get a torrent of ack packets, and can
cause a segfault in the client.  Plus, they don't tell us anything;
those log entries were added when we were making sure that the server
was doing things, and also making sure that the protocol was working
as expected.  They're no longer needed.