Add rest of movements to keyboard driver
Doing the moves is pretty straightforward, and they all use the same
logic, so we'll add the rest of them.  The rotations are not so
simple, and they need some changes in the server to make them work the
way we want, so we'll hold off on them for now.

We'll use a standard-ish WASD cluster for forward-left-back-right
movement.  I'm not sure what the standard up-down keys are, so we'll
use EC for them.  They'll eventually be configurable, so if somebody
doesn't like them, they'll be able to change them.