Simplify stopping movement
Instead of having a separate "stop" command that stops everything
(though we can retain that for now), we'll allow stopping of movement
in one axis, by sending a move with a zero in that given axis.

As things have stood, we always added movements, which resulted in
some pretty insane movements.  Now we'll always *replace*.  This may
yet be not exactly what we want, so there will likely be more changes
as time goes on.