Remove embedded gettext
There's no reason why we need this.  We never build it.
AM_GNU_GETTEXT_VERSION is optional, so removing it just makes sure we
have an external installation.  We don't really care what version
we've got available, because we're not doing anything exotic with it.

We're seeing problems with just removing, so we'll rerun gettextize,
to hopefully fix what's ailing us.  Current version on my dev host is
0.19.8, so there are some changes.