Allow cancelling rotation per axis
We keep track of the translation for all three axes, and can change
any single axis but still retain the other motion.  Doing that with
rotations is a lot more difficult, so we'll split our quaternion out
into Euler angles, zero one out if that's what the request was, and
re-integrate them into a quaternion.

We may eventually want to do something like this for translation.