Complete keyboard rotation controls
We've only been doing yaw up to now, and we should do all three axes
of rotation.  The constants and class members follow the same model as
those for translation.

Since rotation can now be handled per-axis in the server, we don't
need to save anything, and can start and stop rotation per axis.  That
simplifies how we make the action requests for each rotation.