Ensure bundled CuddlyGL works for control objects
When using the bundled CuddlyGL, we're seeing errors when trying to
compile and link the control classes, because they can't find the
CuddlyGL directories.  Default include paths for the control modules
are '.' and '../..', which of course don't get '..'.  Likewise,
default lib paths don't include anyting in our build tree.

Instead of hard-coding anything dumb, we'll use the absolute path of
the installation.

Also, while performing the build the control objects link with the
CuddlyGL library, so we need to make sure it's already built before
trying to build the control objects.