Add active member to game object
We did some testing previously to turn off movement and rotation after
somebody has logged out, by checking the natures lists for the usual
suspects, but we found that it was deathly slow.  Adding a single
boolean variable to keep track of active/inactive has several
advantages:  it's very fast to set and test against; and it's a very
small addition, which can be handled entirely interally to the object.

This allows us to check whether we should be trying to move by
querying (among other things) a single bool.